Made the weapon system actually like a weapon system, and not just a test. Can add weapons normally now. Added melee weapons.

+++ src/animation.hpp
@@ -60,7 +60,7 @@ public:
 	void setReversed(bool reversed);
 	void setReversed(sf::Sprite* box, bool reversed);
 	bool getReversed();
-	sf::Vector2i getSize();
+	sf::Vector2f getSize();
 	void update ();
 
 	std::string getCurrentAnimation();
